feat: zero-movement UOR-native runtime + content-addressed execution/fusion + V&V
Consolidate hologram onto a single content-addressed buffer pool: each value
lives in one 64-byte-aligned buffer and a slot binds to it, so reuse rebinds,
retention keys by κ-label, and constants are pinned — no per-node tensor
movement. Removes the legacy dual fixed-arena + ContentStore copy paths and the
dead Executor / async_session modules.

- exec: BufferArena is the content-addressed pool (buffer.rs); session walks
  per-node with cheap derive_label reuse keys, witnessed derive_label_witnessed
  at output ports (CA-3); whole-graph memo + sub-graph (SG) reuse; byte-bounded
  generational eviction (SC-3). Per-execute scratch reused; Workspace::split_borrow
  returns stack SplitReads = SmallVec<[&[u8];4]> (no per-dispatch heap).
- backend: content-addressed fusion (matmul→activation epilogue, MatMulActivation);
  op_signature for per-node addressing; codec D_MMA round-trips.
- V&V: external-spec conformance (AS/MA/KC/SC/CA/SG/FU) + production-load perf
  (PV-1..4, GFLOP/s + FLOP/core-cycle + scaling) in CONFORMANCE.md / VERIFICATION.md;
  production + content_reuse benches.

68 test suites green; clippy -D warnings clean; fmt clean; no_std (wasm32 +
thumbv7em) builds.

+++ crates/hologram-bench/benches/matmul.rs
@@ -26,15 +26,15 @@ fn make_arena(slot_bytes: usize) -> BufferArena {
     let slots = vec![
         SlotSpan {
             offset: 0,
-            length: rounded as u32,
+            length: rounded as u64,
         },
         SlotSpan {
-            offset: rounded as u32,
-            length: rounded as u32,
+            offset: rounded as u64,
+            length: rounded as u64,
         },
         SlotSpan {
-            offset: 2 * rounded as u32,
-            length: rounded as u32,
+            offset: 2 * rounded as u64,
+            length: rounded as u64,
         },
     ];
     BufferArena::with_capacity(3 * rounded, slots)
